+++ digipipe/esys/esys/schema/__init__.py
@@ -1,9 +1,6 @@
-import copy
 import os
 
-import json
 import pandas as pd
-
 
 here = os.path.dirname(__file__)